+++ NEWS.md
@@ -1,3 +1,11 @@
+# xmlconvert 0.1.1
+
+## Minor changes
+
+* Added function `xml_to_list()` to properly convert XML data to R lists.
+* Added optional argument `first.records` for function `xml_to_df()` to limit the number of XML data records to be converted into a dataframe (e.g. for testing purposes).
+
+
 # xmlconvert 0.1.0
 
 Initial version.